% Generated by roxygen2 (4.0.2): do not edit by hand
\name{construct_haplotypes}
\alias{construct_haplotypes}
\title{Given a set of sequences, extract the haplotypes.}
\usage{
construct_haplotypes(seq_data, cluster_method = "unique",
  cluster_params = list(NULL))
}
\arguments{
\item{seq_data}{The input data}

\item{cluster_method}{The clustering method to use when constructing the
haplotypes}

\item{cluster_params}{A list of input parameters to the clustering
algorithm.}
}
\value{
A list of objects of class \code{\link{Haplotype-class}}
}
\description{
Various different methods can be used to construct the haplotypes. The
following are currently implemented:
\itemize{
  \item{unique - Each unique sequence is a haplotype}
  \item{single - The entire dataset is a single haplotype}
}
}

