% Generated by roxygen2 (4.0.2): do not edit by hand
\name{construct_haplotypes_single}
\alias{construct_haplotypes_single}
\title{Given a set of sequences, extract the haplotypes using the 'single' approach}
\usage{
construct_haplotypes_single(seq_data, cluster_params, n_header_letters)
}
\arguments{
\item{seq_data}{The input data}

\item{cluster_params}{A list of input parameters to the clustering
algorithm.}

\item{n_header_letters}{The number of letters to take from a FASTA header of
one of the sequences and prepend to the haplotype name.}
}
\value{
A list of objects of class \code{\link{Haplotype-class}}
}
\description{
A single haplotype will be constructed from the entire dataset. Unique
sequences will be extracted and their copies will be found and the haplotype
will be constructed using this information.
}

