% Generated by roxygen2 (4.0.2): do not edit by hand
\name{construct_haplotypes_unique}
\alias{construct_haplotypes_unique}
\title{Given a set of sequences, extract the haplotypes using the 'unique' approach}
\usage{
construct_haplotypes_unique(seq_data, cluster_params, n_header_letters)
}
\arguments{
\item{seq_data}{The input data}

\item{cluster_params}{A list of input parameters to the clustering
algorithm.}

\item{n_header_letters}{The number of letters to take from a FASTA header of
one of the sequences and prepend to the haplotype name.}
}
\value{
A list of objects of class \code{\link{Haplotype-class}}
}
\description{
Each unique sequence will be made a haplotype. The names of all the
identical sequences will be added to the haplotype as well as the number of
copies of this sequence.
}

