% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dist_pwa}
\alias{dist_pwa}
\title{Computes the distances between set of target sequences and a single query
sequence}
\usage{
dist_pwa(target_sequences, query_sequence, substitutionMatrix = NULL)
}
\arguments{
\item{target_sequences}{A BStringSet of target sequences}

\item{query_sequence}{The sequence whose distance from the target sequences
must be computed}
}
\value{
A vector of the distances between each target sequence and the
query sequence.
}
\description{
Computes the distances between set of target sequences and a single query
sequence
}

