% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_all_sequences}
\alias{get_all_sequences}
\alias{get_all_sequences,Haplotype-method}
\title{Returns all the sequences of a haplotype}
\usage{
get_all_sequences(the_haplotype)

\S4method{get_all_sequences}{Haplotype}(the_haplotype)
}
\arguments{
\item{the_haplotype}{The haplotype from which the sequences must be
extracted}
}
\description{
Returns all the sequences of a haplotype
}

