% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_test_AAStringSet}
\alias{get_test_AAStringSet}
\title{A function that returns a test AAStringSet}
\usage{
get_test_AAStringSet(dataset_name = "gag_p24_n128_b001")
}
\arguments{
\item{dataset_name}{The name of the test dataset to return}
}
\description{
A function that returns a test AAStringSet
}

