% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_unique_sequences}
\alias{get_unique_sequences}
\alias{get_unique_sequences,Haplotype-method}
\title{Returns the unique sequences of a haplotype}
\usage{
get_unique_sequences(the_haplotype)

\S4method{get_unique_sequences}{Haplotype}(the_haplotype)
}
\arguments{
\item{the_haplotype}{The haplotype from which the unique sequences must be
extracted}
}
\description{
Returns the unique sequences of a haplotype
}

