% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{number_of_unique_sequences}
\alias{number_of_unique_sequences}
\alias{number_of_unique_sequences,Haplotype-method}
\title{Returns the count of the unique sequences in the haplotype}
\usage{
number_of_unique_sequences(the_haplotype)

\S4method{number_of_unique_sequences}{Haplotype}(the_haplotype)
}
\arguments{
\item{the_haplotype}{The haplotype whose sequences must be counted.}
}
\description{
Returns the count of the unique sequences in the haplotype
}

