% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_seq_data}
\alias{read_seq_data}
\title{Reads in the sequence data}
\usage{
read_seq_data(file_name)
}
\arguments{
\item{file_name}{Name of the fasta file}
}
\description{
This function currently just calls readAAStringSet in the Biostrings
package. See that function for more details.
}
\details{
Must be a valid FASTA file with protein sequence data.
}

