% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{seq_hap_distances}
\alias{seq_hap_distances}
\alias{seq_hap_distances,Haplotype,BStringSet-method}
\title{Computes the distances between query sequences and the members of a
haplotype}
\usage{
seq_hap_distances(haplotype, query_sequences)

\S4method{seq_hap_distances}{Haplotype,BStringSet}(haplotype, query_sequences)
}
\arguments{
\item{haplotype}{The haplotype the sequence must be compared to.}

\item{query_sequences}{The sequences to be compared to the haplotype.}
}
\value{
A matrix of the distances between each query sequence and each
sequence in the haplotype. The haplotype sequences correspond to the rows
and the query sequences correspond to the columns.
}
\description{
Computes the distances between query sequences and the members of a
haplotype
}

