% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{total_number_of_sequences}
\alias{total_number_of_sequences}
\alias{total_number_of_sequences,Haplotype-method}
\title{Returns the count of all sequences in the haplotype}
\usage{
total_number_of_sequences(the_haplotype)

\S4method{total_number_of_sequences}{Haplotype}(the_haplotype)
}
\arguments{
\item{the_haplotype}{The haplotype whose sequences must be counted.}
}
\description{
Returns the count of all sequences in the haplotype
}

