% Generated by roxygen2 (4.0.2): do not edit by hand
\name{construct_haplotypes}
\alias{construct_haplotypes}
\title{Given a set of sequences, extract the haplotypes.}
\usage{
construct_haplotypes(seq_data, cluster_method = "unique",
  cluster_params = list(NULL))
}
\arguments{
\item{seq_data}{The input data}

\item{cluster_method}{The clustering method to use when constructing the
haplotypes}

\item{cluster_params}{A list of input parameters to the clustering
algorithm.}
}
\value{
A list of objects of class \code{\link{Haplotype-class}}
}
\description{
The clustering approach to use is described using the list cluster_params.
}
\details{
The simplest haplotype construction approach is to see each unique sequence
as a haplotype. This is achieved by specifying inputs of cluter_method =
'unique' and cluster_params = list(NULL).

Dealing with non-unique sequences leads to several complications.
See the documentation of \code{\link{Haplotype-class}} for more details on
the storing the non-unique data.

I am currently unsure about the best approach to follow for clustering with
the non-unique data. Just clustering with the unique data will probably
produce different results from clustering when the non-unique data is
included. This is a big problem to deal with in the future.
}

