% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get_unique_sequence_details}
\alias{get_unique_sequence_details}
\alias{get_unique_sequence_details,Haplotype-method}
\title{Returns detailed information about the unique sequences in a haplotype}
\usage{
get_unique_sequence_details(the_haplotype)

\S4method{get_unique_sequence_details}{Haplotype}(the_haplotype)
}
\arguments{
\item{the_haplotype}{The haplotype from which the details must be
extracted}
}
\description{
Returns detailed information about the unique sequences in a haplotype
}

