% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Haplotype-class}
\alias{.Haplotype}
\alias{Haplotype-class}
\title{The data structure that holds a haplotype}
\description{
These are the first thoughts on the design of the data structure:
}
\details{
A haplotype will be defined by specifying a number of input sequences
(possibly only one), a distance measure (can be complex and based on a
'custom' substitution matrix), and a threshold. The threshold will the the
maximum average distance that a sequence can be from all the sequences
initally specified and still be consired part of the haplotype.

For this data structure the following methods will be implemented:
\itemize{
  \item is_member_of()
  \item distance_from()
  \item diversity()
  \item ...
}

By specifying only a single input sequence and a distance of zero, it will
be possible to construct haplotypes that consist of only identical
sequences.

The distance concept is extremely complex. There are currently two
different distance approaches under consideration:
\itemize{
  \item biostrings::stringDist - can use either alignment or edit based
  distances
  \item ape::dist.dna - uses phylogenetic distances based on things like the
  Jukes-Cantor model (and more complex variants)
}

To ensure a flexible approach to distances, write a wrapper which allows the
users to select which distance needs to be used. The wrapper can be extended
over time to include both distance functions as well as others. Start by
using stringDist and allowing the use of a custom substitution matrix.
}

